backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
google-genai>=1.0.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.157.0
gspread==6.1.4
apscheduler==3.10.4
python-multipart==0.0.20
httpx==0.28.1
cryptography==44.0.0
aiosqlite==0.20.0
